# Message parser: optimize recursion limits