… core/src/main/java/eu/neverblink/protoc/java/runtime/LimitedCodedInputStream.java
… jmh/src/main/java/eu/neverblink/jelly/jmh/impl/RdfIriParseFallthrough.java
… jmh/src/main/java/eu/neverblink/jelly/jmh/impl/RdfIriParseTableswitch.java